virtio: Enable initial support
This commit adds the Virtio network driver's support in the PCI
device discovery process. This means that when we scan the PCI
bus for devices in pci::init(), any Virtio network cards will be
configured, enabled, and started, with an interrupt handler and
a companion kernel thread started. This will kick off the DHCP
address discovery process, hopefully allocating us an IP address
and discovering gateway and DNS server details.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>