net: Add support for initial workloads after DHCP negotiation
Firely is inherently a networked OS. There's no point starting the
initial workload (what in Unix OSs is the init process) without
being able to use the network.

This adds the ability to resume (typically, start) one or more
threads when a DHCP configuration is negotiated. The threads are
resumed at most once, with the set of initial workloads cleared
once invoked.

Note that it's best to configure the initial workloads before any
network interfaces are initialised, to avoid a race between the
initial workload configuration and the DHCP negotiation process.
DHCP should lose the race, but it's best to play safe.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>