mem: Implement translating virtual to physical memory ranges
This is more involved than translating physical to virtual memory,
as we can't just use the physical memory mapping. Furthermore, it
can be useful to translate a contiguous region of memory. Again,
for physical to virtual translation this is simple, as a contiguous
physical region will result in a contiguous virtual region, but that
is not necessarily true in the other direction.

This introduces a physical buffer type to represent a contiguous
physical memory region. We then implement virt_to_phys_addrs, using
a page table to translate the addresses, producing one or more
regions.

I've included some basic tests to check the behaviour, using the
new debug page table helper.

Note that it would be nice to minimise the number of buffers returned
by merging contiguous physical regions. It's a little fiddly, so it
can wait for another time.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>