virtio: Use virt_to_phys_addrs to make entropy more usable
Previously, we required the user of a virtio entropy device to use
a buffer in the PHYSICAL_MEMORY region, as it made retrieving the
underlying physical memory region simple to determine.

Now that we have virt_to_phys_addrs, we can use that to get the
physical memory buffer(s), no matter where the virtual buffer
resides. We still special case memory in PHYSICAL_MEMORY, for
performance reasons.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>