pci: Moved PCI device driver logic into pci module
Previously, we had a function in the drivers module that optionally
returned a device driver. This worked well, but was PCI-specific
and felt poorly structured.

This commit moves the logic into the PCI module, while also moving
to a slice of device drivers and then using an iterator to find
the first driver that supports the device.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>