thread: Add a suspend function like time::sleep without a timer
As the above suggests, this adds a new function, which can be
called to put the current thread to sleep and witch to the next
runnable thread. The behaviour is identical to time::sleep, but
we don't create a timer to wake the thread back up. The thread
can be re-awoken by calling scheduler::resume or ThreadId.resume.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>