net: Create interfaces in network, not the network driver
Previously, the network driver would construct a smoltcp Interface
and pass that to our network stack's register_interface. This works
fine, but it feels odd for the network stack not to own the interface
and its configuration.

This commit fixes that by moving the interface creation into the
network stack, in the renamed add_interface function.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>