net: Add initial UDP support
This gives us the start of our UDP support. The API is a little
different from Berkeley sockets. Everything starts with a Config
structure, which is used to configure connectivity. This is done
by starting with a default config, then using methods to make any
desired changes. These helper methods enforce constraints on the
configuration changes to ensure only valid configurations can be
represented.

Once the config is ready, open is used to open the port, specifying
the local address. Once the port is open, it can be used to send
and receive packets.

Limitations:

- Currently, only the first network interface is supported.
- There is no support for non-blocking APIs yet.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>