net: Add initial TCP support
This gives us the start of our TCP support. The API is a little
different from Berkeley sockets. The client and server support
is provided through separate APIs, each leading to a Connection.

Client support starts with a DialConfig structure, which is used
to configure the behaviour of new outbound connections. This is
done by starting with a default config, then using methods to make
any desired changes. These helper methods enforce constraints on
the configuration changes to ensure only valid configurations can
be represented.

Once the config is ready, dial is used to open a connection,
specifying the peer's address. Once the connection is open, it
can be used to send and receive streams of bytes.

Server support starts with a ListenConfig structure, which is
used to configure the behaviour of receiving pending connections.
This is done by starting with a default config, then using methods
to make any desired changes. These helper methods enforce constraints
on the configuration changes to ensure only valid configurations
can be represented.

Once the config is read, listen is used to wait for new inbound
connections. Calling accept on the listener is used to accept a new
connection. Once the connection is open, it can be used to send and
receive streams of bytes. The listener can be closed to prevent any
more connections from being accepted, but doing so does not affect
any connections that have already been accepted.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>