gdt: Switch over to CPU-local GDT and TSS once we can
We already have code to set up the global descriptor table and the
task state segment, which is one of the kernel's first tasks. This
works fine, but it has two limitations. Firstly, the TSS is not
mutable, which prevents us from changing any fields. In particular,
we can't change the ring 0 stack pointer, which is necessary to run
code in ring 3. Secondly, we can't yet have a separate instance of
the GDT or TSS for each CPU, which prevents us from running more
than one CPU.

To solve this, we treat our initial GDT and TSS as just a bootstrap.
When we set up CPU-local data, we initialise and switch to a second,
per-CPU GDT TSS pair. This will allow us to have a separate instance
for each CPU, and to safely allow mutable access to both. Finally,
as we would only ever access our own instance, we don't need to
worry about locks or 'thread safety'.

In this commit, we just replicate the GDT and TSS we set up in the
bootstrap. There are no other changes.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>