qemu: Run on an emulated Icelake server CPU
The QEMU CPU is fine for a lot of purposes, but it misses some
useful feature support. This commit tells QEMU to emulate the
Intel Icelake server platform, which has more features.

Note, this may cause several logs starting with:

    qemu-system-x86_64: warning: TCG doesn't support feature:

I've patched the warning out in my copy of Qemu. If you want to
do the same, look for target/i386/cpu.c, mark_unavailable_features.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>