mem: Change StackBounds from inclusive to exclusive
Currently, the StackBounds type contains the first address inside
the range and the first address beyond the range. This has been
fine so far, and was quite convenient, as it kept both addresses
aligned. The fact that the end address was beyond the stack bounds
did not cause problems, as we used it with the push_stack function,
which ecrements the address before writing to it. This means we
immediately decrement the address from the first address outside
the bounds to the last aligned address within the bounds.

The reason this needs to change is that we will want to be able to
allocate userspace stacks at the end of user memory. This means
that the first address after the stack will be non-canonical. If
we store such an address in the CPU, we will trigger an exception.
This commit changes that so that we store the first and the last
addresses within the stack bounds. To account for the fact that
the end address will now be unaligned, we have a small chunk of
code where we prepare the stack, where we subtract 7 from the
address, reproducing the last aligned address in the stack bounds.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>