cpu-local: Add code and data segments for ring 3
To be able to run code in ring 3, we have to have code and stack
segment selectors. This commit defines those in the GDT and the
IA32_SYSENTER_CS MSR. We enforce the invariant that the two selectors
must be sequential, with the stack selector after the code selector.
This is required for sysexit to work, as their values are implied.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>