thread: Add an optional ring 0 interrupt stack pointer
Until now, we've always used a ring 0 interrupt stack pointer of 0,
by leaving the TSS's privilege stack table empty. This means that
when an interrupt occurs, the stack pointer will be left unmodified
by default, continuing to use the existing thread's stack. The only
exception to this is the double fault handler, which has been set
to use the first entry in the interrupt stack table.

To be able to run code in ring 3, we need to be able to set another
stack pointer in the ring 0 stack pointer field. For kernel threads,
this is unnecessary. For now, we just initialise this additional
stack with a stack pointer of 0 in kernel threads. We now update
the privilege stack table whenever we switch thread, using the new
per-CPU TSS. For now, that updates the value to 0 every time, but
if we start creating user threads with a non-zero interrupt stack
address, then the TSS will be updated accordingly.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>