thread: Added an assembly function to start a new user thread
When we start a new kernel thread, we push its entry point onto
the stack, then again with the start_kernel_thread function. This
calls start_kernel_thread when the thread is first executed using
scheduler::switch. This is a tiny assembly function that re-enables
interrupts, clears the stack base pointer, pops the entry point off
the stack, and calls it. This approach is fine for kernel threads,
but doesn't change protection ring and therefore can't be used to
start user threads.

The equivalent functionality for user threads is very similar. We
push the stack pointer onto the stack before the entry point, so
that start_user_thread has access to both. Rather than enabling
interrupts ourselves, we put the user mode entry point and the
stack pointer into RDX and RCX, respectively, and call SYSEXIT.
This enables interrupts, changes protection ring, and changes our
code and stack segments.

With this function, we now have most of the pieces necessary to
start user threads.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>