thread: Add initial support for user threads
This brings everything together to allow user threads to be started.
We allocate a stack (in userspace, rather than in the kernel stack
area) and map it as user-accessible. We then allocate a kernel stack
for handling interrupts. Finally, we prepare the user stack, setting
it up like we do for kernel threads, but using start_user_thread.

This gives us the ability to start user threads, although there's
fairly little those threads can do yet. The next step is to start
implementing syscalls.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>