gdt: Remove the segment selector for GS
It turns out that we don't need to have a segment selector for GS
to be able to use the GSBase MSR, as discussed in rust-osdev/x86_64#333.
I suspect this is because we now have a kernel data segment selector
for SS, which we didn't when I first added support for CPU-local
data. Removing the extra selector frees up a slot in the GDT.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>