thread: Switch from sysexit to sysret for starting user threads
Previously, we used sysexit to 'return' to ring 3 when we start a
new user thread. This works fine, but it places constraints on the
number of GDT entries we use (and their order) and is not supported
in AMD CPUs.

The differences between the two for this use case are:

- We store the RFLAGS in R11, rather than inheriting the current values.
- We provide the RIP in RCX, rather than RDX.
- We inherit the current RSP, rather than setting it in RCX.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>