cpu-local: Set up segment selectors for SYSCALL/SYSRET
Now that we're not using sysexit any more, we can remove our
constant for the IA32_SYSENTER_CS model-specific register.

sysret supports returning to userspace in 32-bit compatibility
or 64-bit long mode. To do so, we need to have both 32-bit
and 64-bit user code segments.

Previously, we used `debug_assert` to enforce the constraints
between the different segment selectors. With syscall/sysret,
this is checked for us in `Star::write`.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>