thread: Store the interrupt stack as StackBounds, not VirtAddr
Previously, we stored user threads' interrupt stack as the virtual
address at the end of the stack bounds (or the null address for
kernel threads). This works fine, but it prevents us from freeing
the stack when the thread exits.

Now, we use an Option<StackBounds> (just like the primary stack),
allowing a more obvious representation of the data and giving us
the information we need to clean up afterwards.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>