thread: Add a syscall stack and preserve syscall and user stacks
Although we interrupts are disabled when handling syscalls, it's
always possible that they will be enabled by the handler, or that
a CPU exception will take place. We don't want to risk stack
corruption, so we allocate a separate kernel stack for the syscall
handler. We will switch to the syscall stack in the syscall handler,
which will be written in assembly. That means we have few tricks at
our disposal. To make stack switching easy, we store the thread's
syscall stack pointer and the saved user stack pointer in the
CPU-local data.

When we switch thread, we might be handling a syscall, in which case
our user stack is saved in the CPU-local data. To preserve this, we
save it into the thread structure and then restore it from the new
thread.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>