shared: Added syscalls crate with partial Firefly ABI
This commit introduces the shared folder, for code shared between
the kernel and userspace. The first example of this is the new
firefly_syscalls crate, which includes constants for each syscall
implemented by the kernel, along with the set of errors that can
be returned by a syscall.

This also makes the repo a workspace, so we add a top-level Cargo.toml
for the workspace, moving the kernel's Cargo.lock into the workspace
root.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>