syscalls: Added initial implementation of syscall handling
This commit gives us the funcationality to set up the syscall
handling, by writing the user and kernel segment selectors to
the IA32_STAR MSR, the address of the syscall handler function
to the IA32_LSTAR MSR, the RFLAGS mask to the IA32_FMASK MSR,
and enable the SYSCALL and SYSRET instructions in the IA32_EFER
MSR. We call the init function that does the above in
kernel::init.

With syscalls set up, user threads can now use the SYSCALL
instruction to call into the kernel. For this, we use the same
ABI as Linux:

- RAX: Syscall number
- RDI: Argument 1
- RSI: Argument 2
- RDX: Argument 3
- R10: Argument 4
- R8:  Argument 5
- R9:  Argument 6

- The user values in RCX and R11 are destroyed by SYSCALL.
- The user values in RAX and RDX are not preserved.
- All other user values are preserved by the kernel.

- RAX: Returned value
- RDX: Returned error

This includes an implementation for the first syscall; ExitThread,
which exits the current thread. If any other syscall is made, its
arguments are printed and BadSyscall is returned.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>