thread: Rename the thread's id to global_id to indicate uniqueness
When we introduce processes, threads will then have two ids. The
existing, globally unique id (used to index into the set of thread
data), and a new per-process id. We will only expose this second
id to userspace. It needs to be clear which we mean, to reduce the
chance of the wrong value being used. To help with this, I've renamed
the existing id to global_id.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>