thread: Do not free user stacks as kernel stacks
When we only had kernel threads, we freed a dropped threads' stack
by passing it to the list of unused kernel stacks. These could
then be used by future threads. Now that we have user stacks, we
must ensure that we only add kernel stacks to the list of unused
kernel stacks. For now, that means user stacks remain allocated,
leaking memory. We can fix that once we have process support.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>