scheduler: Move the preemption logic into the scheduler
Previously, we would act out the preemption logic in the PIT
interrupt handler. This involves decrementing the current thread's
time slice and returning if it remained greater than zero. If the
time slice was less than or equal to zero, we would reset it to
the default value and switch thread.

Functionally, this is fine, but it would be nice if all of the
scheduling logic was in the scheduler, as it would give better
separation of concerns. That's what this commit does. It moves
the existing code into a new `preempt` function in the scheduler
and makes the Thread's tick method private (now that it doesn't
need to be visible to the time module).

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>