storage: Switch from top-level functions to an iterator
Previously, we had a top-level function for each method in the
Device trait. Calling each function would invoke the corresponding
method on the first storage device registered. This worked fine,
but it required us to duplicate all of the functions in the Device
trait.

This replaces that functionality by providing a single function,
allowing the caller to have their callback invoked on each registered
device.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>