filesystem: Added types for file permissions, type, and info
This is just initial types so that they can be used elsewhere. Of
note is the FileInfo type. This highlights some of Firefly's
idiosyncrasies. There is currently support for only files and
directories. Secondly, files are not owned by a particular
identity. As a result, there is a single set of permissions,
which apply to the bearer of the corresponding file handle.

Finally, directories are simpler than in some other OSs. In
Firefly, they have size 0 and cannot have the execute permission.
The read permission is sufficient to explore a directory.

Firefly also has a strict definition for behaviour on a file
that does not have the read permission. In this case, only the
file's name and permissions can be discovered. Its size and
type are both 0.

Note that the file info does not yet contain any timestamps,
as Firefly does not yet supply wall time in any meaningful
sense. This will change in future.

This also defines the filesystem separator as /, like in
UNIXes.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>