tar: Added basic support for reading TAR archives
This adds the ability to read common TAR archives, iterating
through their contents. More advanced archives, such as those
with sparse files, may not be read correctly. Some irregular
files, such as symbolic links are also ignored.

I've included some of the simple test cases from the Go standard
library, but replacing the MD5 checksums with SHA-256, as we
already use the `sha2` crate.

Adding support for more complex TAR files would be nice but is
not high priority. There is no need to implement functionality
to write TAR archives in the kernel.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>