bazel: Add a macro to build bare metal Rust toolchains
To be able to build the kernel, we will need additional Rust
toolchains. This is because the standard toolchain can only
target the host platform. To build the kernel, we need to target
a platform with no OS.

This adds the x86_64_rust_toolchain macro, which can be used to
create a Rust toolchain from a set of compiled standard library
archives and an LLVM target configuration (as a file containing
the JSON structure). The standard library needs to have been
built using the same target configuration.

The reason that this is a macro and not a rule is that we may
have to instantiate multiple Rust toolchains, such as one for
the bootloader and another for the kernel. This is something we
should fix in the long term, but that would require adopting
and forking the bootloader.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>