bazel: Add a macro to build bare metal C++ toolchains
As described in 02690fc, we need to build custom toolchains to
build the kernel for a platform with no OS. Ideally, we would
only need to build Rust toolchains. Sadly, rules_rust uses the
C++ toolchain to perform linking. This means we also need to
configure a corresponding C++ toolchain when we cross-compile.
For more details, see John Millikin's blog post on cross-compiling
Rust here: https://john-millikin.com/notes-on-cross-compiling-rust.

The approach we've taken here is mostly fine, but it does have two
drawbacks. The first is that the toolchain will only run on Linux
on x86_64. It would be nice to support other hosts, but this is
much less easy to do with C++ than with Rust.

The second issue is that we rely on Clang and Ld.LLD from the host
filesystem, rather than building them with Bazel. The proper
approach would be to build them purely within Bazel, which would
also make it easier to cross-compile from more host platforms.

Both of these issues would be nice to fix, but I don't have the
time to investigate them right now.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>