bazel: Build crates core, alloc, and compiler-builtins from source
To be able to cross-compile Rust code, we need to have a copy of
the standard library, built for the target platform. This commit
adds a macro to do just that.

Note that we carefully name the generated rules so that the crates'
names all start with libX. If we don't do that, the tooling will
complain. It's also important that the names don't contain "std",
or rules_rust gets confused and angry. This is annoying, as "no_std"
would be a very obvious name, but so be it.

As documented in the code, this currently uses a genrule, invoking
rules_rust's copy of rustc, rather than using rust_library. This is
because we'd need to use a config transition to ensure that these
crates are always built using the host toolchain. Without that, if
we try to use these crates in a toolchain for a custom target (and
doing so is the entire reason for making them), then there would be
a dependency cycle where the toolchain would depend on these crates
and building the crates would depend on the toolchain.

It shouldn't be too hard to fix, but it's a future problem for now.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>