bazel: Build cross-compiling Rust/C++ toolchains for the kernel
Now that we can access the kernel's target platform, we can build
the standard library for it, construct Rust and C++ toolchains
using the standard library, and bundle them together with a platform.

These will currently require several command line options to enable,
but we can use a config transition to do that for us.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>