bazel: Set up a crates universe so we can use third-party crates
To be able to use third-party crates in rules_rust, we have to set
up a crates universe. This handles fetching the crates, dependency
version resolution, and automatically generating BUILD files.

Note that we use the crates_universe_bootstrap to build the resolver,
which is then used for fetching crates. Setting up the bootstrap
for the first time is quite noisy, as it uses Cargo under the hood
and the resolver currently has a lot of (mostly transitive) dependencies.
This should get better over time.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>