bazel: Added build files for crates we manage directly
Some crates can't be managed sufficiently using the crates
universe. This is mainly because of cases where default features
or optional dependencies pull in std, which we can't use with the
kernel.

The BUILD files used here were generally created automatically and
then modified as necessary to avoid introducing std into the build.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>