bazel: Add a config transition for the kernel
Currently, to use our custom cross-compiling toolchains, we would
have to use several command-line options. This works, but it's
tedious and error prone. Here, we introduce a wrapper rule that
builds on a rust_binary, attaching an outgoing edge transition, as
documented here: https://docs.bazel.build/versions/main/skylark/config.html#user-defined-transitions.

This transition enables the command-line options for us, just for
a single rust_binary and its dependencies. Not only does this handle
the command-line options for us, but it allows us to build multiple
targets, each with different configurations if necessary.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>