bazel: Build the kernel using Bazel
This small commit makes a big difference. Now that we have the
cross-compiling toolchains set up, and all of our dependencies
ready, we can now use a rust_library to build the kernel library
and a rust_binary to build the kernel binary.

The kernel can now be built with `bazel build //kernel`.

Having done so, we can build and run the bootable image by
manually invoking bootimage runner, like `cargo run` does:

	CARGO_MANIFEST_DIR=$PWD/kernel bootimage runner bazel-bin/kernel/kernel

Note that we avoid building the kernel binary directly by default,
by applying the 'manual' tag. This is to ensure `bazel build //...`
doesn't try to build the kernel with the host toolchain. Instead,
we target the wrapper rule that uses the config transition to
switch to the cross-compiling toolchain.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>