bazel: Add a cross-compiling toolchain and transition for bootloader
The bootloader binary uses its own (similar but subtly different)
target platform configuration. As a result, it needs its own Rust
toolchain for that configuration. Similarly, we will need another
wrapper rule to introduce a config transition.

In future, it would be preferable to use a single target for both
the bootloader and the kernel, but that would require forking and
taking ownership of the bootloader, which I'd rather not do yet.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>