bazel: Build the bootloader binary
Now that we can build the kernel binary, we want to build a complete
bootable image. That means building the bootloader binary, which
links in the kernel binary into a single ELF executable. There are
two main stages to this:

1. Run the bootloader's build script to repackage the kernel.
2. Build the bootloader's binary crate, linking in the kernel.

The first stage is somewhat tedious. The build script performs two
tasks here. The first is to identify the kernel's configuration of
the bootloader from its Cargo.toml and produce bootloader_config.rs
accordingly. This covers details like the kernel's stack address
and the virtual offset at which physical memory should be mapped.
The second task is to repackage the kernel binary (stripping debug
symbols, renaming certain symbols, and so on) into a linkable
archive using LLVM tools. As a Cargo build script, it prints two
lines to stdout, instructing Cargo to add the directory containing
the kernel archive to the link search path and to link in the archive.

The second task should be covered entirely by rules_rust's normal
cargo_build_script rule, as it's the kind of functionality build
scripts were designed for. Unfortunately, while it detects the extra
link search line and applies it correctly to the rust_binary, it does
not seem to detect and add the link line. The first task is ignored
completely, as there's currently no way to declare in a cargo_build_script
that the build script produces files used by subsequent rules.

As a result, we build the build script using a rust_binary target
and invoke it using a genrule, so we can capture bootloader_config.rs
and the kernel archive. These are then used as dependencies in the
bootloader binary. The problem with _this_ approach is that there's
no way to extract the parent directory of the kernel archive when
passing it to the binary as to rustc_flags to add to the link search
paths.

The result of the above is that we have to build and run the build
scripts twice! Once in a proper cargo_build_script to add the search
path and once in a rust_binary + genrule to add bootloader_config.rs
and to (manually) add the kernel archive as a link target. It's a bit
ridiculous that we have to do it twice, so finding a way to drop to
one method would be very welcome.

Having resolved all that, building the actual bootloader binary as
a rust_binary is pretty trivial. As with the kernel binary, we use
the manual tag to ensure the bootloader is only built via the config
transition so we don't try to build it with the host toolchain when
using `bazel build //...` and similar.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>