bazel: Convert the bootable ELF binary into a raw disk image
Now that we can build the bootloader binary, linking in the kernel,
we have an ELF binary representing the bootable OS. However, this
is easier to use with Qemu if we transform it into a disk image.
This consists of turning the ELF binary into a raw binary (stripping
the ELF headers and metadata) and padding the binary to an integral
number of 512-byte disk sectors.

This is generally pretty simple, and the code is mostly copied
(with much appreciation) from github.com/rust-osdev/bootimage at
dae6e75abfd60dfe66d46a1566840b5595b53b69. There are just two small
complications:

1. Bazel makes inputs from other targets non-writable.
2. We have to ensure the second stage of the bootloader stays small.

Bazel tries to ensure that a target cannot modify the files it
receives from the targets it depends on. This is to ensure that
everything stays reproducible and predictable and is entirely
sensible. However, this means that the bootable ELF binary is
passed to bootimage with only read and execute permissions.
Sadly, llvm-objcopy copies file permissions when it copies a file,
so the bootable disk image is created non-writable. This then
prevents the bootimage from extending it to the next smallest
disk sector boundary. To solve this, we make a copy of the input
file, ensuring we inlcude writable permissions on the copy.

The second issue is that we have to ensure the bootloader after
stage 1 fits within 127 512-byte disk sectors (63 KiB). Any
larger than this and the BIOS will refuse to copy it using real
mode's int13h functionality. More details on this issue are
available here: https://github.com/rust-osdev/bootloader/issues/205

This issue should not currently occur, but when it does it can be
hard to diagnose. As a result, we use GDB to check that the build
will work before we strip the ELF metadata. This was previously
causing an issue, due to the default configuration for a Rust
toolchain not optimising by default, resulting in the bootloader's
dependencies pushing it well over the 127-sector limit.

With all this done, we can now produce a raw disk image that could
be booted by Qemu!

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>