bazel: Added a build target to run Firefly in Qemu
Just as we configure Cargo to boot Firefly in Qemu with `cargo run`,
this allows us to boot the Firefly using `bazel build //:qemu`.

As in the bootloader (see e609a0c), we have to deal with the fact
that Bazel makes dependency files non-writable to their dependants.
This is perfectly sensible and shouldn't cause any problems, except
Qemu refuses to boot a disk image that isn't writable. Like before,
we solve this by making a copy of our input, making it writable,
and then using that. It's annoying and shouldn't be necessary, but
so be it.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>