bazel: Skip the bootloader size check if GDB isn't installed
While the check that the bootloader is not too large is useful
when things go wrong, it brings no value when things go right.
It shouldn't be necessary to have GDB installed, particularly
for what is just a helper for when a corner case happens.

This commit changes the check to only take place when GDB has
been installed.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>