bitmap: Moved from the kernel crate into its own, named bitmap_index
This commit moves the bitmap module into its own library crate. This
became ambiguous with the bitmap module in kernel::memory::pmm, so
the new crate has also been renamed to bitmap_index.

This should be the first of many modules being extracted into separate
crates. It makes the code easier to manage in the long run and makes
testing considerably easier.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>