pci: Made the PCI module stateless
Previously, the PCI module maintained a vector of device drivers,
passing each identified device to the first driver that claimed
to support it and maintaining a vector of unclaimed devices.

It would be nice to have less state in these kinds of modules,
centralising state in the kernel's core. As a result, this commit
does just that. Now, the `init` function has been renamed to `scan`
and it returns a vector of identified devices. This means that the
set of unclaimed devices can be removed without replacement.

With the driver discovery performed by the caller, there's no need
to keep the `DeviceDriver` or `DeviceDriverSupport` types or the
`DEVICE_DRIVERS` array in the PCI module. These have been moved up
into the drivers module and renamed to `PciDriver`, `PciDriverCheck`,
and `PCI_DRIVERS`, respectively.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>