mmio: Require the caller of Region::map to provide the mapper
Previously, we would use the kernel's PML4 directly when mapping a
new MMIO region. In the longer term, we will want to be able to
customise this, in case we need to map into an individual process's
address space.

In this commit, we instead require the caller to provide the mapper
we use. For now, we just recreate the old behaviour, using kernel_pml4
unconditionally and requiring the mapper to be an OffsetPageTable.
In future, we should probably change it to be anything implementing
the Mapper trait.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>