bugfix: Allocate the right amount of CPU-local memory for each CPU
Previously, we accidentally allocated only enough memory for one
CPU id (rounded up to the next page), rather than the full set of
CPU data. This hadn't caused any problems because the CPU data is
currently less than one page.

This was not a security issue, as the CPU data growing beyond the
size of one page would have resulted in an immediate page fault
while writing the data to memory.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>