gdt: Rewrite the GDT/TSS handling code, export to its own crate
This commit combines two changes. Firstly, the GDT and TSS handling
code in gdt.rs and multitasking/cpu_local.rs has been moved into a
new crate called segmentation. Secondly, the code has been rewritten
to be more robust and correct. The new design abstracts away the
implementation details and correctly enforces lifetime constraints
on the relevant pointers.

Previously, we had two similar but different implementations of the
GDT/TSS handling. Firstly, the global bootstrap in gdt.rs got the
kernel up and running until we had the heap allocator in place.
Once that is done, it hands off to the second implementation in
multitasking/cpu_local.rs. This second version works on a per-CPU
basis, storing the relevant data in the CPU-local data, so that
each CPU could modify its TSS independently (which is necessary to
handle each user thread's kernel interrupt stack). Having two
implementations was tedious enough, but their differences were a
bit of a time bomb.

This commit produces a single implementation, which stores no state
locally. Instead, it allows the caller to store the segment data as
required, then call methods to invoke the relevant functionality.
Notably, we use Pin to ensure that the TSS and GDT cannot be moved
accidentally while loaded into the CPU, which would cause a subtle
and hard-to-debug issue.

Now, the bootstrap code in lib.rs prepares its segment data on the
stack before swapping to the CPU-local version, which is effectively
on the heap. We ensure that segment data is not currently active
when dropped to ensure nothing weird happens.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>