memory: Moved layout constants into a new memlayout crate
This creates a new memlayout crate, containing the address and
address range constants that were previously in memory/constants.rs
and exposed in the memory module. These constants were widely-used,
so having them in a separate crate will make it easier to move
other modules into separate crates, without them depending on the
kernel crate.

Note that the `phys_to_virt_addr` function has moved with it into
the memlayout crate, as it's a static calculation, rather than
being based on the state of the running system (as is the case
with `virt_to_phys_addrs`).

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>