pit: Moved timer handling orchestration into kernel
This is an odd combination of changes but works quite nicely.
Previously, the ticker module in time handled three distinct
duties:

1. Implementing a monotonic ticker.
2. Configuring the Programmable Interrupt Timer's frequency.
3. Providing an interrupt handler for the PIT, which interacted with the scheduler.

It doesn't make a huge amount of sense for all three of those
to live in the same module, particularly given the disproportionate
impact on the time module's dependency graph.

With this change, the ticker module retains duties 1 and 2,
but loses the third, thus removing its interactions with the
interrupts and scheduler subsystems. Instead, the kernel module
brings the three together.

To make this slightly nicer, I've added an Irq::PIT constant
for the PIT's default IRQ number.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>